removed CUDA files. Add Makefile. Renaming

+++ Executor.cpp
@@ -2,11 +2,11 @@
 #include <future>
 #include <stdexcept>
 
-// Однопоточное выполнение
+// One-thread
 class SingleThreadExecutor : public Executor {
 public:
-    std::vector<int> run(const std::vector<int>& data, std::function<int(int)> fn) override {
-        std::vector<int> result;
+    std::vector<float> run(const std::vector<float>& data, std::function<float(float)> fn) override {
+        std::vector<float> result;
         for (auto& item : data) {
             result.push_back(fn(item));
         }
@@ -14,16 +14,16 @@ public:
     }
 };
 
-// Многопоточное выполнение
+// Multi-thread
 class MultiThreadExecutor : public Executor {
 public:
-    std::vector<int> run(const std::vector<int>& data, std::function<int(int)> fn) override {
-        std::vector<std::future<int>> futures;
+    std::vector<float> run(const std::vector<float>& data, std::function<float(float)> fn) override {
+        std::vector<std::future<float>> futures;
         for (auto& item : data) {
             futures.push_back(std::async(std::launch::async, fn, item));
         }
 
-        std::vector<int> result;
+        std::vector<float> result;
         for (auto& future : futures) {
             result.push_back(future.get());
         }
@@ -31,13 +31,17 @@ public:
     }
 };
 
-// Определение метода фабрики
+// Add your methods here
+// ...
+
+// Fabric method definition
 std::unique_ptr<Executor> RunFactory::create(const std::string& method) {
     if (method == "single") {
         return std::make_unique<SingleThreadExecutor>();
     } else if (method == "multi") {
         return std::make_unique<MultiThreadExecutor>();
     } else {
+        // Add your methods here
         throw std::invalid_argument("Unknown method: " + method);
     }
 }
